Arch: Decouple sample registry, harden RT topology, and optimize kernel
- Abstract SampleRegistry into a trait to decouple DNA from audio-core.
- Replace dynamic collections in GraphTopology with fixed-size arrays for RT-safety.
- Wrap SampleMetadata in Arc for zero-allocation O(1) metadata transfers.
- Enhance ProcessorFactory with type-aware capabilities and metadata.
- Optimize StandardKernel loop to reduce sub-block fragmentation.
- Resolve BUG-07 debt regarding command bundle expansion.
- Standardize indices as u32 for stable IPC serialization.

+++ crates/audio-core/src/processors/graph/telemetry.rs
@@ -32,8 +32,8 @@ impl GraphTelemetry {
             let mut node_peak = if offset == 0 { 0.0f32 } else { f32::from_bits(self.peak_levels[n_idx].load(Ordering::Relaxed)) };
 
             for o_idx in 0..routing.output_count {
-                let v_out = routing.output_indices.get(o_idx).copied().unwrap_or(0).min(crate::MAX_NODES - 1);
-                let p_idx = topo.virtual_to_physical.get(v_out).copied().unwrap_or(0).min(crate::MAX_NODES - 1);
+                let v_out = routing.output_indices.get(o_idx).copied().unwrap_or(0).min(crate::MAX_NODES as u32 - 1) as usize;
+                let p_idx = topo.virtual_to_physical.get(v_out).copied().unwrap_or(0).min(crate::MAX_NODES as u32 - 1) as usize;
                 let data = &buffers[p_idx].data[offset..offset + num_samples];
 
                 use wide::*;
